requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
cv2-tools==2.4.0
docopt==0.6.2
EasyProcess==1.1
entrypoint2==1.1
idna==3.4
ImageHash==4.3.1
imutils==0.5.4
Js2Py==0.74
mouse==0.7.1
MouseInfo==0.1.3
mss==7.0.1
np==1.0.2
numpy==1.24.1
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pipwin==0.5.2
pyauto==0.0.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyPrind==2.11.3
PyRect==0.2.0
pyscreenshot==3.0
PyScreeze==0.1.28
pySmartDL==1.3.4
pytesseract==0.3.10
python-constraint==1.4.0
pytweening==1.0.4
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
requests==2.28.2
scipy==1.10.0
six==1.16.0
soupsieve==2.3.2.post1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
win-api==0.0.0
